Refactor Depends to a dependent struct.
